+++ lib/configuration/configuration.ex
@@ -17,6 +17,11 @@ defmodule Interceptor.Configuration do
     get_interceptor_module_function_for(mfa_to_intercept, interception_type)
   end
 
+  def get_interceptor_module_function_for({module, function, nil = _arity} = to_intercept, interception_type) do
+    # functions with arity 0 without parens have their arity as nil
+    get_interceptor_module_function_for({module, function, 0}, interception_type)
+  end
+
   def get_interceptor_module_function_for({module, function, _arity} = to_intercept, interception_type) do
     interception_configuration = get_configuration(module)
     configuration = interception_configuration[to_intercept]