+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Known issues
 
-* intercepted functions without arguments that don't have `()` isn't working (e.g. `def foo, do: 123`)
+- None at this time :)
 
 # Changelog for v0.5.2